[Resolve #65] Change /api/pricecheck/filestatuses/users/id to /api/pricecheck/users/id/filestatuses
