+++ src/main/java/oleksandrdiachenko/pricechecker/controller/FileStatusesRestController.java
@@ -4,9 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import oleksandrdiachenko.pricechecker.controller.exception.EntityNotFoundException;
 import oleksandrdiachenko.pricechecker.model.entity.FileStatus;
-import oleksandrdiachenko.pricechecker.model.entity.User;
 import oleksandrdiachenko.pricechecker.service.FileStatusService;
-import oleksandrdiachenko.pricechecker.service.UserService;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.http.ResponseEntity;
@@ -15,11 +13,8 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import static org.apache.commons.lang3.StringUtils.EMPTY;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @RestController
 @Validated
@@ -29,22 +24,7 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 class FileStatusesRestController {
 
     private final FileStatusService fileStatusService;
-    private final UserService userService;
     private final FileStatusesModelAssembler assembler;
-
-    @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
-    @GetMapping("/users/{userId}")
-    public CollectionModel<EntityModel<FileStatus>> getFileStatusesByUserId(@PathVariable long userId) {
-        User user = userService.findById(userId)
-                .orElseThrow(() -> new EntityNotFoundException("file status", userId));
-
-        List<EntityModel<FileStatus>> employees = fileStatusService.findByUserId(user).stream()
-                .map(assembler::toModel)
-                .collect(Collectors.toList());
-
-        return new CollectionModel<>(employees,
-                linkTo(methodOn(FileStatusesRestController.class).getFileStatusesByUserId(userId)).withSelfRel());
-    }
 
     @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
     @GetMapping("/{id}")
