+++ src/main/java/oleksandrdiachenko/pricechecker/controller/UserRestController.java
@@ -3,7 +3,9 @@ package oleksandrdiachenko.pricechecker.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import oleksandrdiachenko.pricechecker.controller.exception.EntityNotFoundException;
+import oleksandrdiachenko.pricechecker.model.entity.FileStatus;
 import oleksandrdiachenko.pricechecker.model.entity.User;
+import oleksandrdiachenko.pricechecker.service.FileStatusService;
 import oleksandrdiachenko.pricechecker.service.UserService;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
@@ -15,6 +17,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @RestController
 @Validated
@@ -24,7 +30,9 @@ import java.util.List;
 public class UserRestController {
 
     private final UserService userService;
-    private final UserModelAssembler assembler;
+    private final UserModelAssembler userModelAssembler;
+    private final FileStatusesModelAssembler fileStatusesModelAssembler;
+    private final FileStatusService fileStatusService;
 
 
     @PreAuthorize("hasRole('MODERATOR') or hasRole('ADMIN')")
@@ -32,13 +40,27 @@ public class UserRestController {
     public EntityModel<User> getUserById(@PathVariable long id) {
         User user = userService.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("user", id));
-        return assembler.toModel(user);
+        return userModelAssembler.toModel(user);
     }
 
     @PreAuthorize("hasRole('MODERATOR') or hasRole('ADMIN')")
     @GetMapping()
     public CollectionModel<EntityModel<User>> getAllUsers() {
         List<User> fileStatus = userService.findAll();
-        return assembler.toCollectionModel(fileStatus);
+        return userModelAssembler.toCollectionModel(fileStatus);
+    }
+
+    @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
+    @GetMapping("/{userId}/filestatuses")
+    public CollectionModel<EntityModel<FileStatus>> getFileStatusesByUserId(@PathVariable long userId) {
+        User user = userService.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("file status", userId));
+
+        List<EntityModel<FileStatus>> employees = fileStatusService.findByUserId(user).stream()
+                .map(fileStatusesModelAssembler::toModel)
+                .collect(Collectors.toList());
+
+        return new CollectionModel<>(employees,
+                linkTo(methodOn(UserRestController.class).getFileStatusesByUserId(userId)).withSelfRel());
     }
 }
